Move template selection button
Position the template selection button next to the existing button in the Facebook preview component for improved layout and accessibility.
[skip gpt_engineer]
